Extract common SSH host logic into a separate module
